+++ Changelog.md
@@ -20,7 +20,7 @@ OK - Prohibited words must not be case sensitive.
 2.0 OK
 - Put incoming process into a queue. So system will not be overloaded with many simultaneous instances. (now incoming torrents)
 - Mediaworflow.py has been included into TRWorkflow.py
-- 
+- Read user home folder and put logs in $HOME/.TRWorkflow/logs
 
 
 Improvements: